Migrate Fiber tests to ZIO Tests (#1647)
