+++ core-tests/shared/src/test/scala/zio/BaseCrossPlatformSpec.scala
@@ -16,7 +16,7 @@ import zio.system.System
 import scala.concurrent.duration.{ Duration, _ }
 import scala.concurrent.{ ExecutionContext, Future }
 
-abstract class BaseCrossPlatformSpec extends Specification with DefaultRuntime {
+abstract class BaseCrossPlatformSpec extends Specification with DefaultRuntime with LatchOps {
 
   implicit val ec: ExecutionContext = ExecutionContext.Implicits.global
 
@@ -38,19 +38,6 @@ abstract class BaseCrossPlatformSpec extends Specification with DefaultRuntime {
             }
       } yield res
     }
-
-  def withLatch[R, E, A](f: UIO[Unit] => ZIO[R, E, A]): ZIO[R, E, A] =
-    Promise.make[Nothing, Unit] >>= (latch => f(latch.succeed(()).unit) <* latch.await)
-
-  def withLatch[R, E, A](f: (UIO[Unit], UIO[Unit]) => ZIO[R, E, A]): ZIO[R, E, A] =
-    for {
-      ref   <- Ref.make(true)
-      latch <- Promise.make[Nothing, Unit]
-      a <- f(latch.succeed(()).unit, ZIO.uninterruptibleMask { restore =>
-            ref.set(false) *> restore(latch.await)
-          })
-      _ <- UIO.whenM(ref.get)(latch.await)
-    } yield a
 
   implicit def zioAsExecution[A: AsResult, R >: Clock with Console with System with Random, E]
     : AsExecution[ZIO[R, E, A]] =
